+++ src/ui/pages/CreateRoom.jsx
@@ -2,11 +2,17 @@ import { useNavigate } from "react-router-dom";
 import PrimaryButton from "../components/PrimaryButton";
 import { ROOMS_LIST } from "../../routes/routes_consts";
 
-
 export default function CreateRoom() {
-
   const navigate = useNavigate();
-  return <div>CreateRoom
-    <PrimaryButton text='Back to rooms' onClick={(e) => navigate(ROOMS_LIST)} />
-  </div>;
+
+  function handleCreateRoom() {
+    navigate(ROOMS_LIST);
+  }
+
+  return (
+    <div>
+      CreateRoom
+      <PrimaryButton text="Back to rooms" onClick={handleCreateRoom} />
+    </div>
+  );
 }
