+++ src/ui/pages/Error.jsx
@@ -1,5 +1,4 @@
 export const Error = ({ error }) => {
-  console.log("test");
   return (
     <div style={{ textAlign: "center", padding: "2rem" }}>
       <h1>Something went wrong</h1>
